Fix bug in cmake of boots library
The file is deleted, after it was observed that several communities have done so. This helped to correct the error. The file is rewritten, and can be compiled in Visual studio 2022. You should test, in other versions, or in any case, it can be reverted if necessary. https://github.com/AshamaneProject/AshamaneCore/commit/e5b12a76dcd3348ff5ddc3e79da4d4a9f6837351 where AshamaneProject deletes the file.